Module Zeroinstall_solver

module S : sig ... end

Some useful abstract module types.

module Make : functor (Input : S.SOLVER_INPUT) -> sig ... end

Select a compatible set of components to run a program. See Zeroinstall.Solver for the instantiation of this functor on the actual 0install types.

module Diagnostics : functor (Result : S.SOLVER_RESULT) -> sig ... end

Explaining why a solve failed or gave an unexpected answer.

module Sat : sig ... end

The low-level SAT solver.