Module Zeroinstall_solver.Diagnostics

Explaining why a solve failed or gave an unexpected answer.

Parameters

Result : S.SOLVER_RESULT

Signature

module Note : sig ... end

An item of information to display for a component.

module Component : sig ... end

Information about a single role in the example (failed) selections produced by the solver.

type t = Component.t Result.RoleMap.t

An analysis of why the solve failed.

val of_result : Result.t -> t

of_result r is an analysis of failed solver result r. We take the partial solution from the solver and discover, for each component we couldn't select, which constraints caused the candidates to be rejected.

val get_failure_reason : ?⁠verbose:bool -> Result.t -> string

get_failure_reason r analyses r with of_result and formats the analysis as a string.