Module type S.SOLVER_RESULT

module Input : SOLVER_INPUT

include SELECTIONS with module Role = Input.Role and type command = Input.command and type dependency = Input.dependency and type command_name = Input.command_name and type dep_info = Input.dep_info and type requirements = Input.requirements

include CORE_MODEL

module Role : sig ... end

type impl

An impl is something that can fill a Role.t (e.g. a particular version of a package).

type command

A command is an entry-point provided by an implementation. Using a command may require extra dependencies (for example, a "test" command might depend on a test runner).

type command_name = private string

An identifier for a command within a role. Note: It might not be necessary to use any commands - we could instead treat the command name as an optional part of the role, and treat each command as a separate impl instead.

type dependency

A dependency indicates that an impl or command requires another role to be filled.

type dep_info = {

dep_role : Role.t;

dep_importance : [ `Essential | `Recommended | `Restricts ];

dep_required_commands : command_name list;

}

type requirements = {

role : Role.t;

command : command_name option;

}

val requires : Role.t -> impl -> dependency list * command_name list

Get an implementation's dependencies.

The dependencies should be ordered with the most important first. The solver will prefer to select the best possible version of an earlier dependency, even if that means selecting a worse version of a later one (restricts_only dependencies are ignored for this).

An implementation or command can also bind to itself. e.g. "test" command that requires its own "run" command. We also return all such required commands.

val dep_info : dependency -> dep_info

val command_requires : Role.t -> command -> dependency list * command_name list

As requires, but for commands.

val get_command : impl -> command_name -> command option

module RoleMap : Stdlib.Map.S with type RoleMap.key = Role.t

type t

val to_map : t -> impl RoleMap.t

val get_selected : Role.t -> t -> impl option

val selected_commands : impl -> command_name list

val requirements : t -> requirements

val unwrap : impl -> Input.impl

val explain : t -> Role.t -> string

Get diagnostics-of-last-resort.