Module Diagnostics.Make

Parameters

Results : Zeroinstall_solver.S.SOLVER_RESULT

Signature

module Model = Results.Input

module RoleMap = Results.RoleMap

val format_role : Stdlib.Format.formatter -> Model.Role.t -> unit

val format_restrictions : Model.restriction list -> string

module Note : sig ... end

module Component : sig ... end

Represents a single interface in the example (failed) selections produced by the solver. It partitions the implementations into good and bad based (initially) on the split from the impl_provider. As we explore the example selections, we further filter the candidates.

type t = Component.t RoleMap.t

val find_component_ex : RoleMap.key -> 'a RoleMap.t -> 'a

val get_dependency_problem : Model.Role.t -> Component.t RoleMap.t -> Model.impl -> Component.rejection_reason option

val examine_dep : Model.Role.t -> Model.impl -> Component.t RoleMap.t -> Model.dependency -> unit

A selected component has dep as a dependency. Use this to explain why some implementations of the required interface were rejected.

val examine_selection : Component.t RoleMap.t -> Model.Role.t -> Component.t -> unit

val examine_extra_restrictions : Component.t RoleMap.t -> unit

val process_root_req : Component.t RoleMap.t -> Model.requirements -> unit

If we wanted a command on the root, add that as a restriction.

exception Found of Model.Role.t * Model.impl * Model.machine_group

Find an implementation which requires a machine group. Use this to explain the rejection of all implementations requiring other groups.

val check_machine_groups : Component.t RoleMap.t -> unit

module Classes : sig ... end

val check_conflict_classes : Component.t RoleMap.t -> unit

For each selected implementation with a conflict class, reject all candidates with the same class.

val of_result : Results.t -> Component.t RoleMap.t

val pp_rolemap : verbose:bool -> Stdlib.Format.formatter -> Component.t RoleMap.t -> unit

val get_failure_reason : ?⁠verbose:bool -> Results.t -> string

Return a message explaining why the solve failed.