Module Make.Component

Represents a single interface in the example (failed) selections produced by the solver. It partitions the implementations into good and bad based (initially) on the split from the impl_provider. As we explore the example selections, we further filter the candidates.

type rejection_reason = [

| `Model_rejection of Model.rejection

| `FailsRestriction of Model.restriction

| `DepFailsRestriction of Model.dependency * Model.restriction

| `MachineGroupConflict of Model.Role.t * Model.impl

| `ClassConflict of Model.Role.t * Model.conflict_class

| `ConflictsRole of Model.Role.t

| `MissingCommand of Model.command_name

| `DiagnosticsFailure of string

]

type reject = Model.impl * rejection_reason

type t = {

role : Model.Role.t;

replacement : Model.Role.t option;

diagnostics : string Stdlib.Lazy.t;

selected_impl : Model.impl option;

selected_commands : Model.command_name list;

orig_good : Model.impl list;

orig_bad : (Model.impl * Model.rejection) list;

mutable good : Model.impl list;

mutable bad : (Model.impl * rejection_reason) list;

mutable notes : Note.t list;

}

val create : role:Model.Role.t -> (Model.role_information * (Model.impl * Model.rejection) list * string list) -> string Stdlib.Lazy.t -> Model.impl option -> Model.command_name list -> t

val note : t -> Note.t -> unit

val notes : t -> Note.t list

val affected_selection : t -> Model.impl -> bool

val filter_impls_ref : note:Note.t option Stdlib.ref -> t -> (Model.impl -> rejection_reason option) -> unit

val filter_impls : ?⁠note:Note.t -> t -> (Model.impl -> rejection_reason option) -> unit

val apply_restrictions : note:Note.t -> t -> Model.restriction list -> unit

val apply_user_restriction : t -> Model.restriction -> unit

val reject_all : t -> rejection_reason -> unit

val replacement : t -> Model.Role.t option

val selected_impl : t -> Model.impl option

val selected_commands : t -> Model.command_name list

val finalise : t -> unit

val pp_reject : Stdlib.Format.formatter -> reject -> unit

val show_rejections : verbose:bool -> Stdlib.Format.formatter -> (Model.impl * rejection_reason) list -> unit

val rejects : t -> (Model.impl * rejection_reason) list * [> `All_unusable | `Conflicts | `No_candidates ]

val pp_candidates : verbose:bool -> Stdlib.Format.formatter -> t -> unit

val pp_notes : Stdlib.Format.formatter -> t -> unit

val pp_outcome : Stdlib.Format.formatter -> t -> unit

val pp : verbose:bool -> Stdlib.Format.formatter -> t -> unit