Module Make.CommandCache

type t = Cache(CommandRoleEntry).t

The cache is used in build_problem, while the clauses are still being added.

module M : sig ... end

type snapshot = CommandRoleEntry.value M.t

Once the problem is built, an immutable snapshot is taken.

val create : unit -> t

val lookup : t -> (CommandRoleEntry.t -> CommandRoleEntry.value * (unit -> unit)) -> CommandRoleEntry.t -> CommandRoleEntry.value

lookup cache make key will look up key in cache. * If not found, create it with value, process = make key, add value to the cache, * and then call process () on it. * make must not be recursive (since the key hasn't been added yet), * but process can be. In other words, make does whatever setup *must* * be done before anyone can use this cache entry, while process does * setup that can be done afterwards.

val snapshot : t -> snapshot

val get : CommandRoleEntry.t -> snapshot -> CommandRoleEntry.value option

val get_exn : CommandRoleEntry.t -> snapshot -> CommandRoleEntry.value

val filter_map : (CommandRoleEntry.t -> 'a -> 'b option) -> 'a M.t -> 'b M.t