Module Make.RoleEntry

include Model.Role

type t

A role that needs to be filled by a single implementation. If two dependencies require the same role then they will both get the same implementation. This could be the package name, although 0install also includes whether you want source code or a binary in the role.

val pp : Stdlib.Format.formatter -> t -> unit

val compare : t -> t -> int

type value = impl_candidates