Class Make.impl_candidates

method get_clause : candidates -> S.at_most_one_clause option

method get_commands : candidates -> Model.command_name -> (S.lit * Model.command) list

Get just those implementations that have a command with this name.

method get_real_vars : candidates -> S.lit list

Get all variables, except dummy_impl (if present)

method get_vars : candidates -> S.lit list

method get_selected : candidates -> (S.lit * Model.impl) option

method get_state : candidates -> decision_state

method partition : candidates -> (Model.impl -> bool) -> S.lit list * S.lit list

Apply test impl to each implementation, partitioning the vars into two lists. Only defined for impl_candidates.