Module Zeroinstall_solver__Solver_core.Make

Parameters

Model : Zeroinstall_solver.S.SOLVER_INPUT

Signature

module SolverData : sig ... end

We attach this data to each SAT variable.

module S : sig ... end

type decision_state =

| Undecided of S.lit

| Selected of Model.dependency list * Model.command_name list

| Unselected

class type candidates = object ... end

class impl_candidates : Model.Role.t -> S.at_most_one_clause option -> (S.lit * Model.impl) list -> Model.impl option -> object ... end

class command_candidates : Model.Role.t -> S.at_most_one_clause option -> (S.lit * Model.command) list -> object ... end

Holds all the commands with a given name within an interface.

module CommandRoleEntry : sig ... end

module RoleEntry : sig ... end

module ImplCache : sig ... end

module CommandCache : sig ... end

module RoleMap = ImplCache.M

type diagnostics = S.lit

val explain : S.lit -> string

type selection = {

impl : Model.impl;
The implementation chosen to fill the role

commands : Model.command_name list;
The commands required

diagnostics : diagnostics;
Extra information useful for diagnostics

}

val add_replaced_by_conflicts : S.t -> ImplCache.snapshot -> (< get_real_vars : S.lit list; .. > * RoleEntry.t) list -> unit

module Machine_group : sig ... end

On multi-arch systems, we can select 32-bit or 64-bit implementations, but not both in the same set of selections.

module Conflict_classes : sig ... end

val process_self_command : S.t -> (('a * 'b) -> < get_vars : S.lit list; .. >) -> S.lit -> 'b -> 'a -> unit

If this binding depends on a command (<executable-in-*>), add that to the problem. *

parameter user_var

indicates when this binding is used *

@param dep_iface

the required interface this binding targets

val process_dep : S.t -> (Model.Role.t -> < partition : (Model.impl -> bool) -> S.lit list * S.lit list; .. >) -> ((Model.command_name * Model.Role.t) -> < get_vars : S.lit list; .. >) -> S.lit -> Model.dependency -> unit

val make_impl_clause : S.t -> dummy_impl:Model.impl option -> (impl_candidates * Model.Role.t) list Stdlib.ref -> Model.Role.t -> impl_candidates * (S.lit * Model.impl) list

val make_commands_clause : S.t -> (Model.Role.t -> < get_commands : 'a -> (S.lit * Model.command) list; .. >) -> (S.lit -> Model.Role.t -> Model.command_name list -> 'b) -> (S.lit -> Model.dependency list -> unit) -> ('a * Model.Role.t) -> command_candidates * (unit -> unit)

val build_problem : Model.requirements -> S.t -> dummy_impl:Model.impl option -> ImplCache.snapshot * CommandCache.snapshot

Starting from root_req, explore all the feeds, commands and implementations we might need, adding * all of them to sat_problem.

module Output : sig ... end

val do_solve : closest_match:bool -> Output.requirements -> Output.t option